fix: backend skipped initial sync for teamless leagues like badminton
